// src/app/models/academics/assessment-groups-search-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface AssessmentGroupsSearchview {
    sortBy: any;
    name: string[];
    sortOrder: any;
    failedRecords: Array<any>;
    pageNumber: any;
    pageSize: any;
}