// src/app/modules/configuration/assessment/models/assessment-groups-view-model-add.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * AssessmentGroupsViewModelAdd
 */
export interface AssessmentGroupsViewModelAdd { 
    /**
     * Gets or Sets Name
     */
    name: any;
}