// src/app/modules/configuration/assessment/models/assessment-groups-view-model.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Gets or Sets AssessmentGroupsViewModel
 */
export interface AssessmentGroupsViewModel { 
    /**
     * Gets or Sets Name
     */
    name?: string;
    id: string;
}